+++ R/VIM-package.R
@@ -4,7 +4,6 @@
 #' @import sp
 #' @import stats
 #' @import methods
-#' @import MASS
 #' @import nnet
 #' @import e1071
 #' @import grid
@@ -20,6 +19,7 @@
 #' @importFrom graphics polygon rect strheight strwidth text title
 #' @importFrom utils capture.output flush.console head
 #' @importFrom ranger ranger importance
+#' @importFrom MASS stepAIC lqs polr rlm
 #' @useDynLib VIM
 NULL